+++ src/Areas/Api/Controllers/BlobsController.cs
@@ -6,6 +6,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
+using NLog;
 using Weavy.Core;
 using Weavy.Core.Models;
 using Weavy.Core.Services;
@@ -20,7 +21,9 @@ namespace Weavy.Areas.Api.Controllers {
     /// </summary>
     [RoutePrefix("api")]
     public class BlobsController : WeavyApiController {
-        
+
+        private static readonly Logger _log = LogManager.GetCurrentClassLogger();
+
         /// <summary>
         /// Uploads new blob(s). Use multipart/form-data for the request format.
         /// After upload the blobs can be used for setting avatars or as references when creating attachments and/or files. 
@@ -34,23 +37,28 @@ namespace Weavy.Areas.Api.Controllers {
                 throw new HttpResponseException(HttpStatusCode.UnsupportedMediaType);
             }
 
-            // write uploaded files to local disk cache
-            var provider = await Request.Content.ReadAsMultipartAsync(new BlobMultipartFormDataRemoteStreamProvider());
+            try {
+                // write uploaded files to local disk cache
+                var provider = await Request.Content.ReadAsMultipartAsync(new BlobMultipartFormDataRemoteStreamProvider());
 
-            // iterate over the uploaded files and store them as blobs in the database
-            List<Blob> blobs = new List<Blob>();
-            foreach (var data in provider.FileData) {
-                var blob = provider.GetBlob(data);
-                if (blob != null) {
-                    try {
+                // iterate over the uploaded files and store them as blobs in the database
+                List<Blob> blobs = new List<Blob>();
+                foreach (var data in provider.FileData) {
+                    var blob = provider.GetBlob(data);
+                    if (blob == null) {
+                        // file was not uploaded to disk, location contains error message (probably not white-listed)
+                        throw new Exception(data.Location);
+                    } else {
                         blob = BlobService.Insert(blob, System.IO.File.OpenRead(data.Location));
                         blobs.Add(blob);
-                    } catch {
-                        ThrowResponseException(HttpStatusCode.InternalServerError, "Failed to upload blobs");
                     }
                 }
+
+                return new ScrollableList<Blob>(blobs, null, null, blobs.Count(), Request.RequestUri);
+            } catch (Exception ex) {
+                _log.Warn(ex.Message);
+                throw new HttpResponseException(Request.CreateErrorResponse(HttpStatusCode.InternalServerError, ex.InnerException?.Message ?? ex.Message));
             }
-            return new ScrollableList<Blob>(blobs, null, null, blobs.Count(), Request.RequestUri);
         }
     }
 }